refactor: 하이라이트 훅을 HighlightEditor 아래로 위치 이동 및 하이라이트 컴포넌트 export 위치 변경
- HighlightEditorContainer를 src/components/index.tsx에서 export 하도록 수정
