+++ frontend/src/components/highlight/components/HighlightEditor/hooks/index.ts
