+++ frontend/src/components/highlight/components/HighlightEditor/hooks/useHighlightRemoverPosition.ts
