+++ frontend/src/components/highlight/components/HighlightEditor/hooks/useHighlightToggleButtonPosition.ts
