+++ frontend/src/components/highlight/components/index.tsx
@@ -1,0 +1,1 @@
+export { default as HighlightEditorContainer } from './HighlightEditorContainer';
