--- frontend/src/components/highlight/index.tsx
@@ -1,1 +1,0 @@
-export { default as HighlightEditorContainer } from './components/HighlightEditorContainer';
