+++ frontend/src/components/index.tsx
@@ -1,3 +1,4 @@
 export * from './layouts';
 export * from './common';
 export * from './error';
+export * from './highlight/components';
